move PriorityThreadPoolExecutor from MoreExecutors where it is a static class to its own class
